Force windows to load sofaguiqtopengl
