+++ applications/sofa/gui/qt/viewer/EngineBackend.h
@@ -49,7 +49,7 @@ public:
     virtual const std::string screenshotName() =0;
     virtual void screenshot(const std::string& filename, int compression_level) =0;
     virtual void setBackgroundImage(helper::io::Image* image) =0;
-
+	
     virtual bool initRecorder(int width, int height, unsigned int framerate, unsigned int bitrate, const std::string& codec="")  =0;
     virtual void endRecorder() =0;
     virtual void addFrameRecorder() =0;
